move curve types back to the ec root mod
